Remove unused pages and clean up code structure
- Deleted the EvaluateWizard component from the review page.
- Removed the TestMistralPage component from the test page.
- Cleaned up the login form by removing unnecessary whitespace.
- Updated navLinks to remove the Test and Review links, and formatted imports for consistency.
