+++ components/login-form.tsx
@@ -161,7 +161,6 @@ export default function LoginForm() {
                   </FormItem>
                 )}
               />
-
               {/* Login Button */}
               <Button
                 type="submit"
